feat(scoring-v2): MosActive Stage 2 — Scoring Engine v2
Adds ScoringPolicy/PolicyVersion v2 alongside legacy ScoringRule v1:
DRAFT/PUBLISHED/RETIRED/CANCELLED version lifecycle, role/level/status/
newcomer/result components, Decimal(12,4) ROUND_HALF_UP calculation,
Event.start_at-based v1/v2 boundary with retroactive-crossing protection,
Person Status historical applicability, immutable calculation snapshots
with exact reversal, and concurrency-safe stable newcomer sequencing.

fix(scoring-v2): persist stable newcomer sequence for confirmed NO_RULE
Participation

assign_scoring_sequence() is now called as the first step of
calculate_participation(), before role/level/tier/result/version
validation, so a confirmed v2 Participation permanently receives its
sequence position even when scoring ends in NO_RULE. Previously the
sequence was only persisted after the full formula succeeded, so a
confirmed-but-incomplete Participation lost its position and a later
confirmation could take its place — inverting historical newcomer order.
Preview (persist_sequence=False) remains non-persistent. Rewrote the
regression test that asserted the old, wrong behavior into two tests
covering persistence, next-participant sequencing, cancellation
stability, and repair/retry reuse. Updated docs/MOSACTIVE-SCORING-V2.md
to match.

Verified: full test_scoring_v2.py (40 passed), the 11-test targeted
historical-integrity selection, the migration 016 collision/backfill
test, ruff/ruff-format/mypy on changed modules, git diff --check.

Migration 016 is included but not yet run against a staging copy of
production data; no production reconciliation has been performed.

+++ backend/src/event_api/activity_schemas.py
@@ -1,4 +1,5 @@
 from datetime import date, datetime
+from decimal import Decimal
 from typing import Annotated, Literal
 from uuid import UUID
 
@@ -134,7 +135,7 @@ class ParticipationCancelRequest(Contract):
 class ManualAdjustmentRequest(Contract):
     person_id: UUID
     season_id: UUID
-    points: int = Field(ge=-1_000_000, le=1_000_000)
+    points: Decimal = Field(ge=-1_000_000, le=1_000_000, decimal_places=4)
     reason: Reason
     request_id: UUID
